requirements.txt
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
openai==1.14.1
python-telegram-bot==13.7
redis==5.0.1
requests==2.31.0
